+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 - `location`
 - File System API
+- Web Cache API
+
+### Changed
+- New `storage` folder to store localStorage and Web cache
 
 ## [0.1.0] - 2025-07-08
 First version